test(wasm): move BUG-163 page-cap tests to tests/query.rs
Chrome headless runs in GitHub Actions hang mid-suite when the lib test
binary exceeds ~14 IndexedDB-heavy tests — even with 120s timeout and
webdriver.json stabilization flags, the accumulated /dev/shm and
renderer state stalls subsequent tests. Firefox completes the same
17-test binary without issue.

The main merge duplicated two tests that already lived in
tests/query.rs (search_request_roundtrip, search_defaults_skip_stored_fields)
and added four BUG-163 page-cap tests that have the same integration-level
scope. Moving all six to tests/query.rs splits them into their own test
binary with a fresh Chrome instance, bringing lib tests back to 11.

Also made WASM_MAX_PAGE_SIZE pub so external tests can reference the cap
directly rather than hardcoding 1_000.
