+++ searchlite-wasm/src/wasm.rs
@@ -51,7 +51,7 @@ type EventHandler = Rc<RefCell<Option<Closure<dyn FnMut(web_sys::Event)>>>>;
 /// `cursor` / `search_after`. `candidate_size` is intentionally not
 /// capped here — core silently clamps it at `MAX_CANDIDATE_SIZE` and HTTP
 /// does the same — so duplicating that as a hard reject would diverge.
-const WASM_MAX_PAGE_SIZE: usize = 1_000;
+pub const WASM_MAX_PAGE_SIZE: usize = 1_000;
 
 thread_local! {
   // Per-thread (per WASM worker) cache of IndexedDB connections.
@@ -2757,210 +2757,6 @@ mod tests {
     };
     let result = reader.search(&request).unwrap();
     assert_eq!(result.hits.len(), 1);
-  }
-
-  #[wasm_bindgen_test]
-  async fn search_request_roundtrip() {
-    let db = unique_db("searchlite-search-request");
-    let schema = Schema::default_text_body();
-    let schema_json = serde_json::to_string(&schema).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "hello wasm" })];
-    let docs_js = serde_wasm_bindgen::to_value(&docs).unwrap();
-    idx.add_documents(docs_js).unwrap();
-    idx.commit().await.unwrap();
-
-    let request = SearchRequest {
-      query: "hello".into(),
-      fields: None,
-      filter: None,
-      limit: 5,
-      from: 0,
-      return_hits: true,
-      candidate_size: None,
-      #[cfg(feature = "vectors")]
-      max_global_vector_candidates: None,
-      sort: vec![],
-      cursor: None,
-      search_after: None,
-      execution: ExecutionStrategy::Wand,
-      bmw_block_size: None,
-      fuzzy: None,
-      track_total_hits: None,
-      #[cfg(feature = "vectors")]
-      vector_query: None,
-
-      #[cfg(feature = "vectors")]
-      vector_filter: None,
-      return_stored: true,
-      highlight_field: None,
-      highlight: None,
-      collapse: None,
-      aggs: BTreeMap::new(),
-      suggest: BTreeMap::new(),
-      rescore: None,
-      explain: false,
-      profile: false,
-    };
-    let request_json = serde_json::to_string(&request).unwrap();
-    let result = idx.search_request(request_json).unwrap();
-    let result_json: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let hits = result_json["hits"].as_array().unwrap();
-    assert_eq!(hits.len(), 1);
-  }
-
-  // BUG-025: a JS caller passing `limit = 4_000_000_000` to `search()` would
-  // grow the result heap until the WebAssembly linear memory aborts with no
-  // recovery path. The validator must surface a `page_too_large`-style error
-  // before any allocation happens.
-  #[wasm_bindgen_test]
-  async fn search_rejects_limit_above_max_page_size() {
-    let db = unique_db("searchlite-bug-025-search");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "hello" })];
-    let docs_js = serde_wasm_bindgen::to_value(&docs).unwrap();
-    idx.add_documents(docs_js).unwrap();
-    idx.commit().await.unwrap();
-
-    let err = idx
-      .search("hello".to_string(), WASM_MAX_PAGE_SIZE + 1, Some(true))
-      .expect_err("limit above WASM_MAX_PAGE_SIZE must be rejected");
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_search_request");
-    assert!(
-      payload.reason.contains("max page size"),
-      "expected 'max page size' message, got {:?}",
-      payload.reason,
-    );
-
-    // The cap itself is allowed.
-    idx
-      .search("hello".to_string(), WASM_MAX_PAGE_SIZE, Some(true))
-      .expect("limit at the cap must still be accepted");
-  }
-
-  // BUG-025: the JSON-decoded `search_request` path must also reject huge
-  // `limit`, `from`, and `from + limit`, rather than allocating a million-slot
-  // top-k heap.
-  #[wasm_bindgen_test]
-  async fn search_request_rejects_oversized_limits() {
-    let db = unique_db("searchlite-bug-025-request");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "hello" })];
-    let docs_js = serde_wasm_bindgen::to_value(&docs).unwrap();
-    idx.add_documents(docs_js).unwrap();
-    idx.commit().await.unwrap();
-
-    let request = serde_json::json!({
-      "query": "hello",
-      "limit": WASM_MAX_PAGE_SIZE + 1,
-      "return_stored": true,
-    });
-    let err = idx
-      .search_request(request.to_string())
-      .expect_err("oversized limit must be rejected");
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_search_request");
-    assert!(
-      payload.reason.contains("max page size"),
-      "expected 'max page size' message, got {:?}",
-      payload.reason,
-    );
-
-    // `from + limit` over the cap is also rejected, even when `limit` alone
-    // fits.
-    let paginated = serde_json::json!({
-      "query": "hello",
-      "from": WASM_MAX_PAGE_SIZE,
-      "limit": 1,
-      "return_stored": true,
-    });
-    let err = idx
-      .search_request(paginated.to_string())
-      .expect_err("from + limit above cap must be rejected");
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_search_request");
-    assert!(
-      payload.reason.contains("max page size"),
-      "expected 'max page size' message, got {:?}",
-      payload.reason,
-    );
-  }
-
-  // BUG-025: the JsValue-decoded entrypoint shares the same allocation hazard
-  // as `search_request`, so it must enforce the same cap.
-  #[wasm_bindgen_test]
-  async fn search_request_value_rejects_oversized_limit() {
-    let db = unique_db("searchlite-bug-025-request-value");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "hello" })];
-    let docs_js = serde_wasm_bindgen::to_value(&docs).unwrap();
-    idx.add_documents(docs_js).unwrap();
-    idx.commit().await.unwrap();
-
-    let request = serde_json::json!({
-      "query": "hello",
-      "limit": WASM_MAX_PAGE_SIZE + 1,
-      "return_stored": true,
-    });
-    let request_js = serde_wasm_bindgen::to_value(&request).unwrap();
-    let err = idx
-      .search_request_value(request_js)
-      .expect_err("oversized limit must be rejected");
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_search_request");
-    assert!(
-      payload.reason.contains("max page size"),
-      "expected 'max page size' message, got {:?}",
-      payload.reason,
-    );
-  }
-
-  // BUG-025 follow-up: aggregation-only / metadata queries that disable
-  // hits don't grow the result heap, so the page-size cap must mirror
-  // core's behavior and skip the `limit`/`from`/`from + limit` checks
-  // when `return_hits = false`. Matching what `IndexReader::search()` and
-  // `searchlite-http`'s `validate_search` already accept.
-  #[wasm_bindgen_test]
-  async fn search_request_skips_page_cap_when_return_hits_false() {
-    let db = unique_db("searchlite-bug-025-no-hits");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "hello" })];
-    let docs_js = serde_wasm_bindgen::to_value(&docs).unwrap();
-    idx.add_documents(docs_js).unwrap();
-    idx.commit().await.unwrap();
-
-    // `from + limit` well above the cap, but `return_hits = false` — must
-    // be accepted, matching `IndexReader::search()`.
-    let request = serde_json::json!({
-      "query": "hello",
-      "limit": WASM_MAX_PAGE_SIZE + 1,
-      "from": WASM_MAX_PAGE_SIZE + 1,
-      "return_hits": false,
-    });
-    idx
-      .search_request(request.to_string())
-      .expect("oversized limit/from must be allowed when return_hits=false");
-  }
-
-  #[wasm_bindgen_test]
-  async fn search_defaults_skip_stored_fields() {
-    let db = unique_db("searchlite-wasm-return-stored-default");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "hello wasm" })];
-    let docs_js = serde_wasm_bindgen::to_value(&docs).unwrap();
-    idx.add_documents(docs_js).unwrap();
-    idx.commit().await.unwrap();
-
-    let result = idx.search("hello".to_string(), 5, None).unwrap();
-    let parsed: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let hit = &parsed["hits"][0];
-    assert!(hit["fields"].is_null());
   }
 
   #[wasm_bindgen_test]
